Create a module for the VAD script
